requirements.txt
bs4==0.0.2
lxml==5.3.0
pillow==10.4.0
pytest==8.3.3
requests==2.32.3
selenium==4.25.0
